+++ ignite/services/chain/testnet.go
@@ -1,0 +1,37 @@
+package chain
+
+import (
+	"context"
+	"os"
+
+	chainconfig "github.com/ignite/cli/v28/ignite/config/chain"
+)
+
+type InPlaceArgs struct {
+	NewChainID         string
+	NewOperatorAddress string
+	PrvKeyValidator    string
+	AccountsToFund     string
+}
+
+func (c Chain) TestnetInPlace(ctx context.Context, args InPlaceArgs) error {
+	commands, err := c.Commands(ctx)
+	if err != nil {
+		return err
+	}
+
+	// make sure that config.yml exists
+	if c.options.ConfigFile != "" {
+		if _, err := os.Stat(c.options.ConfigFile); err != nil {
+			return err
+		}
+	} else if _, err := chainconfig.LocateDefault(c.app.Path); err != nil {
+		return err
+	}
+
+	err = c.InPlace(ctx, commands, args)
+	if err != nil {
+		return err
+	}
+	return nil
+}